+++ src/components/MessageList/GiphyPreviewMessage.tsx
@@ -1,0 +1,46 @@
+import React from 'react';
+
+import { Message } from '../Message/Message';
+
+import type { StreamMessage } from '../../context/ChannelStateContext';
+import type {
+  DefaultAttachmentType,
+  DefaultChannelType,
+  DefaultCommandType,
+  DefaultEventType,
+  DefaultMessageType,
+  DefaultReactionType,
+  DefaultUserType,
+} from '../../types/types';
+
+export type GiphyPreviewMessageProps<
+  At extends DefaultAttachmentType = DefaultAttachmentType,
+  Ch extends DefaultChannelType = DefaultChannelType,
+  Co extends DefaultCommandType = DefaultCommandType,
+  Ev extends DefaultEventType = DefaultEventType,
+  Me extends DefaultMessageType = DefaultMessageType,
+  Re extends DefaultReactionType = DefaultReactionType,
+  Us extends DefaultUserType<Us> = DefaultUserType
+> = {
+  message: StreamMessage<At, Ch, Co, Ev, Me, Re, Us>;
+};
+
+export const GiphyPreviewMessage = <
+  At extends DefaultAttachmentType = DefaultAttachmentType,
+  Ch extends DefaultChannelType = DefaultChannelType,
+  Co extends DefaultCommandType = DefaultCommandType,
+  Ev extends DefaultEventType = DefaultEventType,
+  Me extends DefaultMessageType = DefaultMessageType,
+  Re extends DefaultReactionType = DefaultReactionType,
+  Us extends DefaultUserType<Us> = DefaultUserType
+>(
+  props: GiphyPreviewMessageProps<At, Ch, Co, Ev, Me, Re, Us>,
+) => {
+  const { message } = props;
+
+  return (
+    <div className='giphy-preview-message'>
+      <Message message={message} />
+    </div>
+  );
+};